+++ CHANGELOG.rst
@@ -70,6 +70,7 @@ Fixed
 
 [0.17.10]
 * Update logger statements
+* Added new class logManager for managing repititive logs
 
 [0.17.9]
 * Update TelModel version to 1.18.2